agent/requirements.txt
livekit-agents>=0.10.0
livekit-api>=0.10.0
livekit-plugins-openai>=0.7.0
openai>=1.0.0
python-dotenv>=1.0.0

